refactor(dsh-web-file-preview): 客户端入口按职责拆分（#18 方案 B2）
拆分 909 行的 src/client/index.ts 为 7 个职责模块 + 1 个装配入口：
- state.ts: 状态类型与工厂（apply 内闭包，杜绝模块级全局变量）
- dom.ts: DOM 工具函数（el/ensureStyle/copyPathText/errorView）
- preview.ts: Modal 生命周期（openPreview/closeModal/finalizeSession/onKeyDown）
- text.ts: 文本渲染与 diff（fetchText/renderTabBody/probeDiff/renderDiff）
- image.ts: 图片渲染与灯箱（renderImage/renderBlobImage/openLightbox/closeLightbox）
- intercept.ts: 点击拦截与文件识别（onClickCapture/findFileLink/isPreviewablePath 等）
- wrapper.ts: openPath 收口包装
- index.ts: 仅剩 apply() 装配 + 生命周期布线 + inject/export

设计要点：
- 状态全部收进单一 FilePreviewState 对象，经参数传递，无模块级全局变量
- 按避免循环依赖原则微调函数归属：fileUrl→preview.ts, fileDiffUrl→text.ts,
  upgradePreviewImages/pumpPreviewImages→text.ts（仅被 renderTabBody 调用）
- 每个新文件 ≤ 400 行（wc -l 自证）
- 零 DOM 行为变化，仅搬移代码

Refs #18
